add C10

+++ Just_Lib/ft_convert_base.c
@@ -18,7 +18,7 @@
 
 char	*ft_convert_base(char *nbr, char *base_from, char *base_to);
 void	ft_putnbr_base(int nbr, char *base, char *res, int *i);
-size_t	ft_strlen(const char *str);
+int	ft_strlen(const char *str);
 int		ft_check_base(char *str);
 int		ft_findchr(char *base, char c);
 
@@ -110,7 +110,7 @@ int	ft_findchr(char *base, char c)
 	return (-1);
 }
 
-size_t	ft_strlen(const char *str)
+int	ft_strlen(const char *str)
 {
 	size_t		i;
